fix parallel tests
